% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\docType{data}
\name{Time}
\alias{Time}
\title{Time}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Time
}
\description{
Time
}
\examples{
x <- Time$new()
x$now()
x$now_with_mock_time()
}
\keyword{datasets}
