% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_fuzz.R
\docType{data}
\name{time_fuzz}
\alias{time_fuzz}
\title{time_fuzz}
\value{
an \link{time_fuzz} object
}
\description{
time_fuzz
}
\details{
\strong{Methods}
\describe{
\item{\code{freeze(...)}}{
Freez time
}
\item{\code{scale(...)}}{
Freez time
}
\item{\code{local(...)}}{
Freez time
}
\item{\code{travel(...)}}{
Freez time
}
\item{\code{return(...)}}{
Freez time
}
}
}
\examples{
\dontrun{
x <- time_fuzz$new()
x
x$freeze
x$freeze(x = "2019-01-29", {
  5 + 5
})

x$scale
x$scale({
  5 + 5
})
}
}
\keyword{datasets}
