% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timefuzz-package.R
\docType{package}
\name{timefuzz-package}
\alias{timefuzz-package}
\alias{timefuzz}
\title{Time Travel to Test Time Dependent Code}
\description{
Time Travel to Test Time Dependent Code
}
\author{
Scott Chamberlain \href{mailto:sckott@protonmail.com}{sckott@protonmail.com}
}
\keyword{package}
