% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_stack_item.R
\name{TimeStackItem}
\alias{TimeStackItem}
\title{TimeStackItem}
\description{
TimeStackItem
}
\examples{
\dontrun{
x <- TimeStackItem$new(mock_type = "freeze", date = "2019-02-18")
x
x$time
x$time$time
x$time$year
x$scaling_factor_
x$travel_offset_
x$year()
x$min()
}
}
